# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TaskMaster API
  description: Production-grade Collaborative Task Tracking System RESTful API documentation
  contact:
    name: TaskMaster Engineering Team
    url: https://github.com/taskmaster
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.1.0
servers:
- url: http://localhost
  description: Generated server url
security:
- BearerAuth: []
...
paths:
...
  /api/v1/tasks/{id}:
    get:
      tags:
      - Tasks
      summary: Get task by ID
      operationId: getTaskById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
      security:
      - BearerAuth: []
    put:
      tags:
      - Tasks
      summary: Update task details
      operationId: updateTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
      security:
      - BearerAuth: []
    delete:
      tags:
      - Tasks
      summary: Soft delete a task
      operationId: deleteTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringString"
      security:
      - BearerAuth: []
...
  /api/v1/tasks:
    get:
      tags:
      - Tasks
      summary: Search and filter tasks with multi-field sorting and pagination
      operationId: searchTasks
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - REVIEW
          - COMPLETED
          - ARCHIVED
      - name: priority
        in: query
        required: false
        schema:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - URGENT
      - name: assigneeId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: creatorId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: teamId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: search
        in: query
        required: false
        schema:
          type: string
      - name: dueDateFrom
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: dueDateTo
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: label
        in: query
        required: false
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponsePageResponseTaskResponse"
      security:
      - BearerAuth: []
    post:
      tags:
      - Tasks
      summary: Create a new task
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
      security:
      - BearerAuth: []
...
  /api/v1/tasks/{id}/status:
    patch:
      tags:
      - Tasks
      summary: Transition task lifecycle status
      operationId: updateTaskStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskStatusRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
      security:
      - BearerAuth: []
  /api/v1/tasks/{id}/assign:
    patch:
      tags:
      - Tasks
      summary: Assign or unassign task
      operationId: assignTask
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTaskRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTaskResponse"
      security:
      - BearerAuth: []
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
          enum:
          - USER
          - ADMIN
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
        description:
          type: string
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - URGENT
        dueDate:
          type: string
          format: date-time
        labels:
          type: array
          items:
            type: string
            maxLength: 50
            minLength: 0
          uniqueItems: true
    ApiResponseTaskResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/TaskResponse"
        meta:
          type: object
          additionalProperties:
            type: object
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - REVIEW
          - COMPLETED
          - ARCHIVED
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - URGENT
        dueDate:
          type: string
          format: date-time
        creator:
          $ref: "#/components/schemas/UserResponse"
        assignee:
          $ref: "#/components/schemas/UserResponse"
        teamId:
          type: string
          format: uuid
        labels:
          type: array
          items:
            type: string
          uniqueItems: true
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 0
        description:
          type: string
        priority:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
          - URGENT
        dueDate:
          type: string
          format: date-time
        assigneeId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        labels:
          type: array
          items:
            type: string
            maxLength: 50
            minLength: 0
          uniqueItems: true
      required:
      - title
...
    ApiResponseMapStringString:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: string
        meta:
          type: object
          additionalProperties:
            type: object
...
    UpdateTaskStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum:
          - OPEN
          - IN_PROGRESS
          - REVIEW
          - COMPLETED
          - ARCHIVED
      required:
      - status
    AssignTaskRequest:
      type: object
      properties:
        assigneeId:
          type: string
          format: uuid
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    ApiResponsePageResponseTaskResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/PageResponseTaskResponse"
        meta:
          type: object
          additionalProperties:
            type: object
    PageResponseTaskResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TaskResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
...
  securitySchemes:
    BearerAuth:
      type: http
      description: "Enter your JWT token in the format: Bearer <token>"
      name: BearerAuth
      scheme: bearer
      bearerFormat: JWT